Update TestUniqueExistencePosition.java
I edited this code and included within it the things I removed, why I removed them, and things that I added.

It's still a work in progress, but I fixed some bugs + I got some new insights looking through it that may help to integrate this into larger program.

Please read through comments in it to understand what was changed.